try GCN again

+++ model/temporal/temporal_ext_gcn.py
@@ -95,7 +95,7 @@ class TemporalExtGCN(nn.Module):
     def forward(self, x):
         node_x, edge_idx, edge_attr = x.x, x.edge_index, x.edge_attr
 
-        node_x = node_x
+        node_x = node_x.double()
         edge_idx = edge_idx
         edge_attr = edge_attr
